fix(ci): pass ORCH-0785, ORCH-0863 C7, and append-only gates for Ve1 PR
Resend no-attachment opt-outs, ORCH-0099 backend allowlist on orch-0863,
restore ORCH-0855 personaFork test unchanged and add Ve1 tests in new file.
